Fix the least square problem in matrix.

+++ include/matrix.h
@@ -17,7 +17,7 @@
 typedef struct {
 	int     i;
 	int     j;
-	int     total;
+	long    total;
 	double *element;
 } MATRIX;
 
@@ -34,7 +34,7 @@ MATRIX *matrix_transpose( const MATRIX * );
 MATRIX *matrix_inverse( const MATRIX * );
 
 MATRIX *matrix_assign( MATRIX *, const double, int, int );
-MATRIX *matrix_assign_seq( MATRIX *, const double *, const int );
+MATRIX *matrix_assign_seq( MATRIX *, const double *, const long );
 MATRIX *matrix_assign_row( MATRIX *, const double *, int, const int );
 MATRIX *matrix_assign_col( MATRIX *, const double *, int, const int );
 MATRIX *matrix_assign_diag( MATRIX *, const double *, const int );
@@ -45,7 +45,7 @@ MATRIX *matrix_apply_col( MATRIX *, double (*)( const double ), int );
 MATRIX *matrix_apply_diag( MATRIX *, double (*)( const double ) );
 
 double *matrix_prefill_array( double *, int, ... );
-double *matrix_extract_seq( const MATRIX *, double *, const int );
+double *matrix_extract_seq( const MATRIX *, double *, const long );
 
 double  matrix_determinant( const MATRIX * );
